% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/m13.R
\name{cocluster_accuracy}
\alias{cocluster_accuracy}
\title{Statistic which measures overlap between equal length factors x and y}
\usage{
cocluster_accuracy(x, y)
}
\description{
Statistic which measures overlap between equal length factors x and y
}
