% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/m13.R
\name{plot.TreeStructure}
\alias{plot.TreeStructure}
\title{Plot TreeStructure tree with cluster and partition variables}
\usage{
\method{plot}{TreeStructure}(x, use_ggtree = TRUE, ...)
}
\arguments{
\item{x}{A TreeStructure object}

\item{use_ggtree}{Toggle ggtree or ape plotting behaviour}

\item{...}{Additional arguments passed to ggtree or ape::plot.phylo}
}
\description{
Plot TreeStructure tree with cluster and partition variables
}
