% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/treestructure.R
\docType{package}
\name{treestructure-package}
\alias{treestructure}
\alias{treestructure-package}
\title{treestructure: Detect Population Structure Within Phylogenetic Trees}
\description{
TODO
}
\details{
Includes the *trestruct* function for partitioning a tree. 
This includes methods for printing and plotting. 
Refer to the vignettes for common usage.
}
\keyword{internal}
