% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/treestructure.R
\docType{package}
\name{treestructure-package}
\alias{treestructure}
\alias{treestructure-package}
\title{treestructure: Detect Population Structure Within Phylogenetic Trees}
\description{
This package implements algorithms for detecting population structure from the history of coalescent events recorded in phylogenetic trees. The *trestruct* method classifies each tip and internal node of a tree into disjoint sets characterized by similar coalescent patterns.
}
\details{
Includes the *trestruct* function for partitioning a tree. 
This includes methods for printing and plotting. 
Refer to the vignettes for common usage.
}
\keyword{internal}
