% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/treestructure.R
\docType{package}
\name{treestructure-package}
\alias{treestructure}
\alias{treestructure-package}
\title{treestructure: Detect Population Structure Within Phylogenetic Trees}
\description{
This package implements algorithms for detecting population structure from the history of coalescent events recorded in phylogenetic trees. The *trestruct* method classifies each tip and internal node of a tree into disjoint sets characterized by similar coalescent patterns.
}
\details{
Methods for detecting structure in phylogenies. Includes the *trestruct* function for partitioning a tree and methods for printing and plotting trees with structure. 
Refer to the vignettes for common usage.
}
\references{
Erik Volz, Carsten Wiuf, Yonatan Grad, Simon Frost, Ann Dennis, Xavier Didelot, "Identification of hidden population structure in time-scaled phylogenies", bioRxiv 704528
}
\keyword{internal}
