% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/treestructure-package.R
\docType{package}
\name{treestructure-package}
\alias{treestructure}
\alias{treestructure-package}
\title{treestructure: Detect Population Structure Within Phylogenetic Trees}
\description{
Algorithms for detecting population structure from the history of coalescent events recorded in phylogenetic trees. This method classifies each tip and internal node of a tree into disjoint sets characterized by similar coalescent patterns. The methods are described in Volz, E., Wiuf, C., Grad, Y., Frost, S., Dennis, A., & Didelot, X. (2020) \doi{10.1093/sysbio/syaa009}.
}
\details{
Methods for detecting structure in phylogenies. Includes the *trestruct*
function for partitioning a tree and methods for printing and plotting trees
with structure.
Refer to the vignettes for common usage.
}
\references{
Erik Volz, Wiuf Carsten, Yonatan Grad, Simon Frost, Ann Dennis, Xavier Didelot,
   "Identification of hidden population structure in time-scaled phylogenies",
   (2020); Systematic Biology, 69: 884–896.
}
\seealso{
Useful links:
\itemize{
  \item \url{http://emvolz.github.io/treedater/}
  \item \url{https://github.com/emvolz/treedater}
  \item Report bugs at \url{https://github.com/emvolz/treedater/issues}
}

}
\author{
\strong{Maintainer}: Erik Volz \email{erik@erikvolz.info} (\href{https://orcid.org/0000-0001-6268-8937}{ORCID})

Other contributors:
\itemize{
  \item Fabricia F. Nascimento \email{thednainus@yahoo.com} (\href{https://orcid.org/0000-0001-9426-6835}{ORCID}) [contributor]
  \item Vinicius B. Franceschi \email{vinibfranc@gmail.com} (\href{https://orcid.org/0000-0002-0006-9337}{ORCID}) [contributor]
}

}
\keyword{internal}
