% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/m12.R
\name{trestruct}
\alias{trestruct}
\title{Partition rooted tree into sets of clades}
\usage{
trestruct(tre, minCladeSize = 20, minOverlap = -Inf, nsim = 1000,
  level = 0.001, ...)
}
\arguments{
\item{tre}{A tree of type ape::phylo. Must be rooted and binary.}

\item{minCladeSize}{All clusters within parititon must have at least this many tips.}

\item{minOverlap}{}

\item{nsim}{Number of simulations for comparing clusters within partition}

\item{level}{Significance level for splitting partition to make new cluster}
}
\value{
A TreeStructure object which includes cluster and partitition assignment for each tip of the tree.
}
\description{
Partition rooted tree into sets of clades
}
\examples{
library(ape)
tree <- rcoal(100)
struct <-  trestruct( tree )
}
